+++ decidim-bulletin_board-app/app/commands/enqueue_message.rb
@@ -39,8 +39,12 @@ class EnqueueMessage < Rectify::Command
   attr_reader :client, :signed_data, :job, :pending_message
 
   def valid?
-    client && election && message_id && signed_data && job &&
+    client && election && message_id && signed_data && job && !pending_messages? &&
       (election.authority == client || election.trustees.member?(client))
+  end
+
+  def pending_messages?
+    PendingMessage.enqueued.where(client: client, election: election).exists?
   end
 
   def create_pending_message!