+++ decidim-bulletin_board-ruby/spec/decidim/bulletin_board/client_spec.rb
@@ -88,6 +88,18 @@ module Decidim
         end
       end
 
+      describe "cast_vote_message_id" do
+        let(:election_id) { "test.1" }
+        let(:voter_id) { "voter.1" }
+
+        context "when everything went ok" do
+          it "calls the cast_vote_message_id method and returns message_id" do
+            message_id = subject.cast_vote_message_id(election_id, voter_id)
+            expect(message_id).to eq("decidim-test-authority.test.1.vote.cast+v.voter.1")
+          end
+        end
+      end
+
       describe "cast_vote" do
         let(:election_id) { "test.1" }
         let(:voter_id) { "voter.1" }
@@ -117,6 +129,31 @@ module Decidim
         end
       end
 
+      describe "get_pending_message_status" do
+        let(:message_id) { "decidim-test-authority.test.1.vote.cast+v.voter.1" }
+
+        context "when everything went ok" do
+          before do
+            stub_wisper_publisher("Decidim::BulletinBoard::Voter::GetPendingMessageStatus", :call, :ok, "accepted")
+          end
+
+          it "calls the GetPendingMessageStatus command and returns the result" do
+            get_pending_message_status = subject.get_pending_message_status(message_id)
+            expect(get_pending_message_status).to eq("accepted")
+          end
+        end
+
+        context "when something went wrong" do
+          before do
+            stub_wisper_publisher("Decidim::BulletinBoard::Voter::GetPendingMessageStatus", :call, :error, "Sorry, something went wrong")
+          end
+
+          it "calls the GetPendingMessageStatus command and throws an error" do
+            expect { subject.get_pending_message_status(message_id) }.to raise_error("Sorry, something went wrong")
+          end
+        end
+      end
+
       describe "get_election_status" do
         let(:election_id) { "decidim-test-authority.1" }
 
@@ -126,7 +163,7 @@ module Decidim
           end
 
           it "calls the GetElectionStatus command and returns the result" do
-            election_status = subject.get_status(election_id)
+            election_status = subject.get_election_status(election_id)
             expect(election_status).to eq("key_ceremony")
           end
         end
@@ -137,7 +174,7 @@ module Decidim
           end
 
           it "calls the GetElectionStatus command and throws an error" do
-            expect { subject.get_status(election_id) }.to raise_error("Sorry, something went wrong")
+            expect { subject.get_election_status(election_id) }.to raise_error("Sorry, something went wrong")
           end
         end
       end